refactor: Formatter 클래스 CustomNumberFormatter로 rename 및 파일분리
- CalculatorManager의 nestedType에서 독립된 별개의 클래스로 수정
